Review maintainer scripts
In prerm, don't drop the alternatives going through "active-theme" they are
the responsibility of desktop-base. In postinst, continue to create them
to update them in case the new kali theme provides the file that the
former theme did not.

Fix the check protecteding the call to plymouth-set-default to use a
version that actually existed in kali-themes even though we don't really
have an upgrade case where we wanted to call the command. But it will
make it easier to force a re-run in the future, in case we change the
name of the plymouth theme.