Add the logo-text and logo-text-version variants as symlinks to logo.svg
logo-text is meant to be a version of the logo with the full name
and our logo includes the name already.

logo-text-version is meant to include a version of the release but
we are a rolling release (and the version changes too often) so it
dosn't make sense to hardcode any version here either.