Add a note regarding kali-wallpapers-2021.4
This package is "back from the future". It contains the wallpapers for
Kali version 2021.4 (to be released at the end of 2021). However:

1) This package is already ready
2) Daniel wants to release it now, as the wallpapers it contains were
   already presented to the rest of the world

So we want that the package is pulled in during upgrades, or is part of
new ISOs, and we use Recommends for that.

The wallpapers in there are NOT used by default.

Gbp-Dch: ignore